refactor: reorganize & simply file structure
