+++ renderer/audio.html
@@ -8,7 +8,7 @@
   </style>
 </head>
 <body>
-  <h1><span class="icon icon-note-beamed"></span></h1>
+  <h1>🎵</h1>
   <audio id="audio"></audio>
   <audio id="needle"></audio>
   <script>require('./audio/index.js')</script>
