--- renderer/elements/artwork/styles.js
@@ -1,31 +1,0 @@
-var css = require('csjs-inject')
-
-module.exports = css`
-  .albumCover {
-    position: relative;
-    width: 100%; /* desired width */
-    border-radius: 4px;
-    overflow: hidden;
-  }
-  .albumCover:before {
-    content: '';
-    display: block;
-    padding-top: 100%; /* initial ratio of 1:1*/
-  }
-
-  .albumArt {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-
-    background: #eee;
-    background-size: cover;
-    background-position: center;
-  }
-
-  @media (max-width: 460px) {
-    .albumCover { border-radius: 0 }
-  }
-`
