--- renderer/elements/meta/styles.js
@@ -1,7 +1,0 @@
-var css = require('csjs-inject')
-
-module.exports = css`
-  .meta { margin: 1rem 0 0 }
-  .title, .artist { margin: 0 }
-  .title { font-weight: 600 }
-`
