--- renderer/elements/search/styles.js
@@ -1,12 +1,0 @@
-var css = require('csjs-inject')
-
-module.exports = css`
-  .searchInput {
-    width: 100%;
-    padding: 1px 5px;
-    vertical-align: middle;
-    min-height: auto;
-    margin-right: 1em;
-    height: 24px;
-  }
-`
