--- renderer/lib/artwork-cache.js
@@ -1,3 +1,0 @@
-var cache = {}
-
-module.exports = cache
