+++ renderer/pages/main/index.js
@@ -1,15 +1,15 @@
 var html = require('choo/html')
 var Header = require('../../elements/header')
-var Table = require('../../elements/table')
-var Footer = require('../../elements/footer')
+var Playlist = require('../../elements/playlist')
+var Player = require('../../elements/player')
 
 var styles = require('../styles')
 
 module.exports = main
 
 var header = new Header()
-var table = new Table()
-var footer = new Footer()
+var playlist = new Playlist()
+var player = new Player()
 
 // TODO set up view instance factory
 function main (state, emit) {
@@ -17,9 +17,9 @@ function main (state, emit) {
     <main class="${styles.window}">
       <div class="${styles.grow}">
         ${header.render(state, emit)}
-        ${table.render(state, emit)}
+        ${playlist.render(state, emit)}
       </div>
-      ${footer.render(state, emit)}
+      ${player.render(state, emit)}
     </main>
   `
 }
