+++ renderer/pages/preferences/index.js
@@ -1,8 +1,8 @@
 var html = require('choo/html')
 var Header = require('../../elements/header')
 var button = require('../../elements/button')
-var Footer = require('../../elements/footer')
-var footer = new Footer()
+var Player = require('../../elements/player')
+var player = new Player()
 var formStyles = require('../../elements/form/styles')
 var { app, dialog } = require('electron').remote
 var styles = require('../styles')
@@ -45,7 +45,7 @@ function preferences (state, emit) {
           </div>
         </div>
       </div>
-      ${footer.render(state, emit)}
+      ${player.render(state, emit)}
     </main>
   `
 }
